requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
apscheduler==3.10.4
python-dotenv==1.0.0
streamlit==1.28.2
webdriver-manager==4.0.1